requirements.txt

Flask
tensorflow
pillow
flask-cors
